Refactor to use a sessionManager
