+++ tygron-environment/src/main/java/nl/tudelft/contextproject/tygron/TygronSession.java
@@ -1,11 +1,6 @@
 package nl.tudelft.contextproject.tygron;
 
-import org.json.JSONArray;
 import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
 
 /**
  * TygronSession.
@@ -76,56 +71,21 @@ public class TygronSession {
   public void setSessionId(int newId) {
     this.sessionId = newId;
   }
-
+  
   /**
-   * Return a list of joinable tygron sessions.
-   * 
-   * @return A list of joinable tygron sessions.
+   * Set a new server token.
+   * @param serverToken The server token.
    */
-  public List<TygronSession> getJoinableSessions() {
-    List<TygronSession> returnList = new ArrayList<TygronSession>();
-
-    JSONArray data = apiConnection.callPostEventArray(
-        "services/event/IOServicesEventType/GET_JOINABLE_SESSIONS/", null);
-
-    for (int i = 0; i < data.length(); i++) {
-      JSONObject row = data.getJSONObject(i);
-
-      // Create new session
-      TygronSession localSession = new TygronSession(apiConnection);
-
-      // Set data here
-      localSession.setName(row.getString("name"));
-      localSession.setType(row.getString("sessionType"));
-      localSession.setLanguage(row.getString("sessionType"));
-      localSession.setSessionId(row.getInt("id"));
-
-      // Add to datalist
-      returnList.add(localSession);
-    }
-
-    return returnList;
+  public void setServerToken(String serverToken){
+    this.sessionServerToken = serverToken;
   }
   
   /**
-   * Join a Tygron session and reply whether it was a success or not. 
-   * Data is loaded into memory.
-   * @return whether the join was succesful or not.
+   * Set a client token.
+   * @param clientToken The client token.
    */
-  public boolean joinSession(int slotId) {
-    HashMap<String,String> dataMap = new HashMap<String,String>();
-    dataMap.put("0", slotId + "");          // Param: Integer: Server slot ID
-    dataMap.put("1", "VIEWER"); // Param: AppType: My application type: EDITOR, VIEWER, ADMIN, BEAM 
-    dataMap.put("2", "");                   // Param: String: My client address (optional)
-    dataMap.put("3", "Tygron-API-Agent");   // Param: String: My client computer name (optional)
-    dataMap.put("4", "");                   // Param: UUID: My client token for rejoining (optional)
-    JSONObject data = apiConnection.callPostEventObject(
-        "services/event/IOServicesEventType/JOIN_SESSION/", dataMap);
-    
-    this.sessionClientToken = data.getString("sessionClientToken");
-    this.sessionServerToken = data.getString("serverToken");
-    
-    return true;
+  public void setClientToken(String clientToken){
+    this.sessionClientToken = clientToken;
   }
 
   /**
